LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/window-open-invalid-url-expected.txt


PASS Window.open should throw SyntaxError when an invalid url is passed as the first parameter
